+++ lib/main.dart
@@ -1,7 +1,8 @@
-import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
+import 'package:studdyplanner/todo_detail_dialog.dart';
 import 'package:studdyplanner/todo_service.dart';
+import 'package:studdyplanner/todolist_tile.dart';
 import 'package:table_calendar/table_calendar.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
@@ -76,7 +77,7 @@ class _MyHomePageState extends State<MyHomePage> {
                     context: context,
                     builder: (context) {
                       return AlertDialog(
-                        content: TodoDeailPage(
+                        content: TodoDeailDialog(
                           index: todoService.todoList.length - 1,
                         ),
                       );
@@ -113,11 +114,11 @@ class _MyHomePageState extends State<MyHomePage> {
                         itemBuilder: (context, index) {
                           Todo todo = todoList[index];
                           return isDeleteMode
-                              ? TodoListDeleteCell(
+                              ? TodoListDeleteTile(
                                   todo: todo,
                                   index: index,
                                 )
-                              : TodoListCell(
+                              : TodoListTile(
                                   index: index,
                                   todo: todo,
                                 );
@@ -145,179 +146,6 @@ class _MyHomePageState extends State<MyHomePage> {
                 ),
         );
       },
-    );
-  }
-}
-
-class TodoListCell extends StatelessWidget {
-  const TodoListCell({required this.index, required this.todo});
-  final Todo todo;
-  final int index;
-
-  @override
-  Widget build(BuildContext context) {
-    return ListTile(
-      title: Text(
-        todo.content,
-        maxLines: 3,
-        overflow: TextOverflow.ellipsis,
-      ),
-      onTap: () {
-        showDialog(
-          context: context,
-          builder: (context) {
-            return AlertDialog(
-              content: TodoDeailPage(
-                index: index,
-              ),
-            );
-          },
-        );
-      },
-    );
-  }
-}
-
-class TodoListDeleteCell extends StatefulWidget {
-  const TodoListDeleteCell({
-    super.key,
-    required this.todo,
-    required this.index,
-  });
-  final Todo todo;
-  final int index;
-
-  @override
-  State<TodoListDeleteCell> createState() => _TodoListDeleteCellState();
-}
-
-class _TodoListDeleteCellState extends State<TodoListDeleteCell> {
-  bool? isSeleted = false;
-  @override
-  Widget build(BuildContext context) {
-    TodoService todoService = context.read<TodoService>();
-
-    return CheckboxListTile(
-      title: Text(
-        widget.todo.content,
-        maxLines: 3,
-        overflow: TextOverflow.ellipsis,
-      ),
-      controlAffinity: ListTileControlAffinity.leading,
-      value: isSeleted,
-      onChanged: (bool? value) {
-        setState(() {
-          isSeleted = value;
-          if (value != null && value) {
-            todoService.deleteList.add(widget.todo.id);
-          } else if (value != null && !value) {
-            todoService.deleteList.remove(widget.todo.id);
-          }
-        });
-      },
-    );
-  }
-}
-
-class TodoDeailPage extends StatelessWidget {
-  const TodoDeailPage({required this.index});
-
-  final int index;
-
-  @override
-  Widget build(BuildContext context) {
-    TodoService todoService = context.read<TodoService>();
-    Todo todo = todoService.todoList[index];
-    Size screenSize = MediaQuery.of(context).size;
-    return Container(
-      width: screenSize.width * 0.6,
-      height: screenSize.height * 0.3,
-      decoration: BoxDecoration(
-        borderRadius: BorderRadius.circular(10),
-      ),
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        mainAxisAlignment: MainAxisAlignment.spaceBetween,
-        children: [
-          Padding(
-            padding: const EdgeInsets.all(8.0),
-            child: Container(
-              width: double.infinity,
-              decoration: BoxDecoration(
-                  border: Border(
-                      bottom: BorderSide(
-                color: Colors.grey,
-                width: 1,
-              ))),
-              child: Text(
-                "나의 할일",
-                style: TextStyle(
-                  fontSize: 24,
-                ),
-              ),
-            ),
-          ),
-          Container(
-            constraints: BoxConstraints(
-              maxHeight: screenSize.height * 0.15,
-            ),
-            child: SingleChildScrollView(
-              scrollDirection: Axis.vertical,
-              reverse: true,
-              child: Padding(
-                padding: const EdgeInsets.all(8.0),
-                child: Center(
-                  child: TextField(
-                    onChanged: (value) {
-                      todoService.updateTodo(index: index, content: value);
-                    },
-                    controller: TextEditingController(text: todo.content),
-                    decoration: InputDecoration(
-                      focusedBorder: InputBorder.none,
-                    ),
-                    autofocus: true,
-                    textAlign: TextAlign.center,
-                    maxLines: null,
-                    keyboardType: TextInputType.multiline,
-                    style: TextStyle(
-                      fontSize: 20,
-                    ),
-                  ),
-                ),
-              ),
-            ),
-          ),
-          Row(
-            children: [
-              Expanded(
-                flex: 1,
-                child: TextButton(
-                  onPressed: () {
-                    Navigator.pop(context);
-                  },
-                  child: Text(
-                    "취소",
-                    style: TextStyle(
-                      color: Colors.red,
-                    ),
-                  ),
-                ),
-              ),
-              Expanded(
-                flex: 1,
-                child: TextButton(
-                  onPressed: () {
-                    Navigator.pop(context);
-                  },
-                  child: Text(
-                    "확인",
-                  ),
-                ),
-              ),
-            ],
-          )
-        ],
-      ),
     );
   }
 }
